backend/requirements.txt
google-genai>=0.4.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
websockets>=11.0
